emissivity

Compute land surface emissivity.

emissivity(
  x,
  nonveg = 0.2,
  veg = 0.5,
  enonveg = 0.95,
  eveg = 0.99,
  pveg = FALSE
)

Arguments

x
NDVI raster layer.

nonveg
NDVI threshold value for non-vegetated pixels.

veg
NDVI threshold value for vegetated pixels.

enonveg
Emissivity value for non-vegetated pixels.

eveg
Emissivity value for vegetated pixels.

pveg
Logical; if TRUE, proportion of vegetation is given as an output.

Value

Raster layer containing emissivity values. If pveg = TRUE, a list with emissivity and pveg raster layers.

Details

Computes the per-pixel emissivity based on the modified NDVI threshold method (Sobrino et al., 2008). Threshold values for non-vegetated and vegetated pixels can be defined by the user. Defaults are nonveg <= 0.2 and veg >= 0.5, based on the NDVI values. Emissivity values for non-vegetated and vegetated pixels can be also defined. Defaults are enonveg = 0.95 and eveg = 0.99. pveg is logical. If TRUE, the proportion of vegetation layer is added to the output. Default is FALSE.

References

Sobrino, J. A., Jim?nez-Mu?oz, J. C., S?ria, G., Romaguera, M., Guanter, L., Moreno, J. & Mart?nez, P. (2008). Land surface emissivity retrieval from different VNIR and TIR sensors. IEEE Transactions on Geoscience and Remote Sensing, 46(2): 316-327.

Examples

# For default non-vegetation and vegetation emissivity values
emiss <- emissivity(x, enonveg = 0.95, eveg = 0.99, pveg = TRUE)

#> Error in emissivity(x, enonveg = 0.95, eveg = 0.99, pveg = TRUE): objeto 'x' no encontrado